# [SPARK-10585][SQL] only copy data once when generate unsafe projection