@cloud-fan 2015-09-25
I tried to generate as less code as possible and move the code to real java classes, to debug and maintain it easier.