@davies 2015-10-01
@cloud-fan I did one pass, left a few minor comments. Considering all the combination of data types, This is pretty complicated, could you add more tests to improve the test coverage?