@davies 2015-10-05
LGTM, left some minor comments about coding style.