@tyro89 2015-12-02
I believe there might be an issue here with respect to the buffer growing strategy: https://issues.apache.org/jira/browse/SPARK-12089